Remove toggle arrow from collapsible menu
The toggle arrow is defined by a unicode character, unfortunately
not all fonts have that unicode character mapped. In those cases
the toggle arrow will display as a square.

This update removes the toggle arrow. We could consider adding a
transparent GIF to represent the arrow, though we run into a number
of troubles:
- sizing
- foreground color contrasting against the background
- getting the arrow to look nice and be anti-aliased

Given that a user can visually tell when the collapsible menu
is expanded or collapsed, and learning that it is toggled by clicking
the menu button, it's not ideal but we can perhpas live without
the toggle arrows for now.