Fix battle dialogs appearing behind the battle window.
This update removes the setting: Dialog.ModalExclusionType.APPLICATION_EXCLUDE

When this setting is on, battle dialogs become "associated" with JFrame that
renders the map instead of the battle dialog frame. This means that if you
minimize the map, then the dialog is also minimized. Instead we battle dialogs
to be associated with the battle dialog.

With the 'bad' association we 'fixed' it by having the battle window
always being on top. Usually this meant that the map was rendered, then
the battle window, then any dialogs. 'always on top' has OS specific
behavior and this set us up for a situation where the 'always on top' would
'win out' over the modal dialog being displayed by the battle dialog.

After removing the APPLICATION_EXCLUDE modal type, battle dialogs become
associated with the battle window once again helping ensure that battle
dialogs will always be in front of the battle window, and the battle window
still has always on top so it should always be in front of the map.