Default anonymous login checkbox to true
UX improvement for new users to default the anonymous login
checkbox to true. When seeing a login screen for the first time,
the presence of a password and "create account" button field
is enough to trigger an assumption that the user must register.

The assumption you must register is not true.

To facilitate as much as possible the initial login, default
the anonymous login to true, as brand new users will not
have accounts, nor are they required to have them. So long
as accounts are not required, it's a misleading UX to make
it appear that accounts *are* required.

Note: any selection made to the anonymous login checkbox is
remembered between game sessions. This update is oriented
towards brand new users.