Add distinct callback for websocket terminated (#6236)
Currently we have only one callback whenever a websocket
is closed. This update creates a distinction between
when a websocket is 'terminated' (closed by server)
vs 'closed' (standard close by client).

The 'closed' callback remains just a runnable and continues
to be just for client side cleanup. The 'terminated' callback
is invoked with the disconnect error reason, which for the
case of bans will be the ban message from the server.

If the server gives no message, then we default that message
to a generic "the server disconnected" message.

After this update users are given a notification when they
are disconnected due to ban, before they were given no notification.