Increate codeclimate thresholds
Cognitive complexity of 7 seems to be a more reasonable
threshold where methods are actually quite difficult. We have
had some examples where 6 and 7 are hard to reduce down further.

Method line count of 25 is almost about right, breaking 30 starts
to be more clear cut and reduces noise at that threshold.