Remove "removeChatListener" methods
The remove chat listener methods are there to "prevent" a memory leak but do not
do so. This update removes those methods.

Whenever we remove 'chat listeners' we are always assigning chat to a new object
which removes references to chat. References that chat in turn are holding will
be garbage collected, they do not need to be removed explicitly.

It is also a bit noteworthy we are inconsistent about removing all listeners
each time, and it is also noteworthy we do not remove other references that
chat holds that would by the previous rationale would have also caused
leaks.

The 'add/remove' listener pattern in swing is quite old and was there to
either deal with static state or to enable/disable listeners. It is more
common to instead now allow objects to go out of scope which would make
any attached listener references then also become candidates for GC.