Add lobby right-click menu for moderators: "Show Player Information"
This update adds an equivalent right-click menu feature that was in 1.9
where moderators can click chatter names in lobby and select 'show player
information'.

The player information panel shows a copy-pastable summary of the
players name, IP and system-id and currently two tabs: aliases & bans.

The aliases are obtained from a lobby query and database fetch where we
search the 'access log' table for the player IPs or system-ids matching
the current user and return matching records.

We fetch player aliases limited to the last 'n' days, bans we will
fetch all that we have on record.

Change listing:
- add UI component to show player information
- add two tab UI components to show player ban and alias information in a table
- add a client query to ModeratorChatClient to fetch player information
- add a controller endpoint to service this query, restriced to moderator access
- add a player information module on server side to aggregate and fetch
  player information from database.
- add a player information DAO with JDBI return data types to query database
  for bans and aliases.