Fix TOCTOU when getting images from an image cache.
There exists a 'clear cache' method that can possibly remove
the data in 'images'. If a method call interleaving is just right
then we have a TOCTOU problem where we can detect an image exists,
clear it, then try to retrieve it resulting in a NPE.

This is fixed by getting the image and checking if that is null,
if not we use the image we retrieved.