Convert 'removeGame' call to lobby to be async
If we close a lobby connected game it triggers a 'removeGame' call
to lobby. If that fails with exception, then the shutdown logic
is aborted resulting in a game that stays open.

This update moves the 'removeGame' call to be async so that any
we do not wait for a lobby response to shut down a game and
any exceptions are logged as part of the async call.

Address: https://github.com/triplea-game/triplea/issues/6336