Update error report title to use TripleA class from stack trace
This update changes the preferred title for triplea error report uploads
to look for any exceptions in the reported log record, look for
any exception causes, and if so to look for any TripleA classes
from the stack trace and use that class, method and line number
in the error report title.

For example:
```
2.0.0: PlayerAttachment#getCanTheseUnitsMoveWithoutViolatingStackingLimit:253 - RuntimeException
```

Before we reported the class name where the logging was being done. For the
general uncaught exception handler this was always where we initialized
the general exception handler and was not as useful, for example:
```
NullPointerException - HeadedGameRunner.lambda$initializeClientSettingAndLogging$0
```

The above example after this update would be instead reported as:
```
2.0.0: TripleAFrame#showGame:2266 - NullPointerException
```

Noting that the stack trace looked like this:
```
games.strategy.triplea.ui.TripleAFrame.lambda$showGame$58(TripleAFrame.java:2266)
``

If there is no exception message attached to a logging, then we'll go with old
behavior of reporting the class where the logging was done with logging message.