Improve error message when failing to connect to the lobby
We are seeing cases where connections to lobby can fail with no error message.
When this happens the user is unable to do anything with at least chat
(their chat messages are not sent nor received).

This update improves the error message and informs the user that they should
reconnect.