Fix improved artillery support not being applied
This update fixes improved artillery support. Essentially
the problem is a 'put' and 'add' was replaced with a 'put' and 'put'.
In the first case we put a value of '1' and then added to it, in the
latter (buggy) case we replaced the existing '1' vlaue with another
'1'.

To explain a bit further, support is tracked with three data structures,
- one for units that can receive support
- one for the support rules
- one for units providing support.

Improved artillery support works by double adding artillery units
to the map of units that can provide support.

When removing a custom data structure we replaced a map add
with a map put, causing the data structure for artillery units
to be written twice with a value of '1' rather than adding '1'

Addresses: https://github.com/triplea-game/triplea/issues/6282