Update staging screen alliances button - click to play alliance
Alliances buttons can now be clicked to play that alilance

Changes the behavior of the alliances button on the single player
game staging screen to set the human player to play all factions
of a given alliance.

Previously the button would replicate the current select player
in the row to each faction of the alliance, this behavior was
not very intuitive nor terribly useful.

Secondarily, the tooltip text of 'click to play alliance' button
was set on game load, in which case it always read
"click to [alliance] to human" which could be incorrect if the
current row had been changed. Now the text is more generic
 and always accurate "click to play [alliance]"