Remove lobby disconnect+reconnect notification to hosted games
Removing to prevent a potential message storm if the connection
to lobby is dropped and re-established repeatedly. This was
observed after putting laptop to sleep and then opening again
(result in 20+ some notifications).

It is perhaps better to just let be and have the reconnect
be silent, avoid interrupting a possibly in-progress game.
If the player is on a staging screen, likely they are connected
to lobby chat and will see notifications there when their
chat is disconnected, so user will potentially still be
notified.