Make displayed dice rolls span multiple lines via a WrapLayout.
This change improves the display of dice rolls when there's a lot
of them, by wrapping the dice as opposed to showing them all in
one scrollable line. To accomodate the change, which can increase
the height of the containing components, some UI elements are
restructured to use an outer scrollpane - particularly the combat
and history views.