Rename 'Super-Mod' to 'Admin'
Historically a moderator was called an admin. In 2.0 with
user-roles, the 'admin' role is now called a 'moderator'.
To avoid previous confusion the term 'super-mod' was used
to refer to an admin. This update replaces the 'super-mod'
term with 'admin' (to recall, an admin is someone that
can add and remove moderators).