Suppress option to submerge if subs attacking undefended V3+ transports
If subs are attacking undefended V3 transports, or if after
a round of combat only V3 transports are left, then subs should
not be able to retreat and the transports are auto-destroyed.

This update adds a check to see if we have such a case and if so
skips the subs submerge prompt.

Address: Part (2) of https://github.com/triplea-game/triplea/issues/4688