Smoke-testing, add test that all current maps can be parsed
- Add list of all current map XMLs (map-list.txt)
- Updates smoke-testing/build.gradle to download all XMLs from map-list.txt
- Adds a test to smoke-testing that loads each XML and verifies
  each one can be parsed.

Goal: this test is to ensure we do not have any map parsing regressions
and that we can continue parsing any existing XMLs. This will help
ensure that the engine is not updated to break parsing and secondarily
that maps available for download will not fail on parsing.

Note: we can potentially see a problem if a map becomes broken in repository.
If that happens then we can comment out the map XML in map-list.txt until
it is fixed. On the other hand if we have a non-backward compatible update
we can do similar while updating maps (this highlights the problem we have
with non-backward compatible map updates).