Update 0 vs 0 stalemate check to take into account dice rolls
Instead of only looking at just attack power for determining
if both defense and offense have zero power, also look at the
number of dice rolls available and consider the battle
a stalemate if both sides either have no rolls or no power.
This is done simply by multiplying the power and rolls together,
if either is zero the product will be zero.