Fix NPE for routeDescription being null
Was able to repro via attempting non-combat moves in edit
mode on map 'Middle Earth Battle for Arda' while 'Perform Move'
was not selected. After some effort was able to trigger the code
for calculating a move without a routeDescription set.

This update adds a simple NPE check to avoid dereferncing a
null routeDescription. The fule cost is later used to render a
movement route, considering there might be none this should be fine.