Remove end of script travis failure catch block
So far it looks like we will consistently report a failure if any step
fails. This update removes a travis catch block that triggers on failure
and also reported a failure. This should simplify the messaging we get
in any 'master branch build failed' issue and list a failure just once.

A risk from this update is if we get a failure in an unexpected place
that is not explicitly handled.