Locally runnable prod deployment script
Unify the name of the deployment scripts in 'infrastructure'.
After this update we now have:
  - run_ansible_vagrant
  - run_ansible_prerelease
  - run_ansible_production
Each is similar and can be run locally. The travis scripts
are now updated to set up any preconditions, notably copying
the ansible vault secret variable to a vault file and then
kicking off the respective 'run_ansible' script.