Infrastructure: Fix path to http-server jar file
The jar file for http-server is extracted to a 'bin/' folder.
This update corrects the path of run scripts to refer to that
folder and updates the 'latest deployment' variant to copy
a jar file to that same location. We did not catch this before as
prerelease builds are from source and pushes artifacts from local
to remote, while version specific builds (for production) are
downloaded and then unzipped.