Add an explicit log message when handling unified messenger remote exceptions
Using just an exception message for a log message can make it hard to correlate
an observed message to the location in code that generated that message. Notably
using an exception message and the message can be variable and secondarily
there won't be any code that has an exact matching string.

This update adds an explicit message prefix so we can have more context
if we ever see an exception handled by unified messenger.