Fix bots to do shallow parsing of zipped maps
This update changes:
1. bots to do 'shallow' parsing of zipped map files.
   A shallow parse fetches only a limited amount of data.
2. Add logging to bots to print which map is being loaded.

Bots do 'shallow' parsing of expanded maps (folder) but were doing
full parsing of zipped maps when fetching the list of available games.
Full parsing is slower and evaluates the full Map XML converting
all nodes into a GameData object. This is unnecessary work when
we only need the game name and version from the XML.