Execute websocket pings with 3 retries & 3 second backoff
If the ping send fails, we'll get disconnected (disconnect is
at 60s, we send a ping every 45s). We could increase the ping
frequency at the cost of greater load. Instead we use a simple
retry to give any failures a few more attempts up until
pretty close to the 60s mark. If we are unable to send a ping
by then, then we would expect the websocket connection to
server to be closed. When that connection is closed, we will
show a pop-up to the user that the connection was lost
(so there is no need to inform a user when a ping fails,
just when the connection is lost will we notify the user).

.