Merge database roles (#6643)
* Clean up ansible playbook tags, remove tags that are not heavily used.

* Remove redundant roles in bots group, admin and firewall are already applied in the 'all' block

* Merge lobby_database tasks into postgres role tasks

* Simplify apt module usage, can accept a list of packages without an explicit vars

* Combine lobby_database and postgres roles

The postgres role was developed to be a generic postgres database.
The lobby_database installed the lobby_db database onto the postgres
database. The idea was if we had more application databases then
we'd have more roles. This way you could re-use the postgres role
to install postgres but have other roles to install more databases.

This update merges the two roles and converts the application
database set to a datastructure: "name, user, password". By defining
these triplets, the postgres role will now create databases
for each triplet defined with the corresponding database name, user
and password. This way we can just use the postgres database role
which can then create the application databases without additional roles.

Also of note, the check for a password being set is combined, if that
fails then we use the shell command to create a user password. This is
useful to combine as the result variable becomes a list and is difficult
to loop over in a subsequent command.

* Remove settings from playbook that can be specified in ansible.cfg instead