Throttle code climate similar code threshold to 80, avoid warnings on imports
Code climate is detecting similar code warnings in imports with masses ranging
from 55 to 75. Generally, 4 to 5 lines of similar imports is not something
we want to change. This update throttles the similar code threshold to be
less sensitive, to a value that will omit these false positives (a value of 80)