Script up summing of code-climate report files
The cc-report script only works when run from the root of the source directory.
Apparently the --prefix flag that is meant to do this does not work.
To generate a report we CD to each source directory of each subproject
that generates an XML file, run the coverage formatter, and then
as a last step we sum up each coverage file and upload.