Restore deprecated map property costPu
Some players might have ancient maps, the error from the missing property
is non-obvious to recover from. Furthermore, there is misleading error
messaging that the map is not present and a prompt to download, which
might not even result in a fix and then another error when the
game becomes the default selection.

This problem speaks to being able to version maps for when they are
not forward compatible and to be able to download engine specific versions.