Conditional coverage report upload based on env variable presence
Travis environment variables are only present for branch builds from
the main repo, not forks. This update makes coverage uploads conditional
on whether these variables are not present. Otherwise we get an error
for variable not defined when building from forks.

Second, the 'CC_TEST_REPORTER_ID' env variable can be defined directly
in Travis so we do not need to re-assign, this has been made implicit.