Update PR template with release note block
The format & syntax for release note is documented in wiki.
The release comment goes in between the special release
note comment block, eg:
`UPDATE|Add release note magic block to PR template`

Also in this update, uncomment sections so they are always include and
its okay if they are left blank. This makes the documentation and
manual steps slightly less when submitting a PR.