BugFix: Lookup if map is available by game name
The list of available map folders and zip is not as reliable
as the list of game names. We are seeing an odd issue where
'World War II Revised' is not on the available folder and zip
list but is available by game name. By looking up save
games by game name, we can avoid this problem.