Begin extracting nodeBB forum http interface
Starting with the 'getUserId' API, extract from NodeBbForumPoster
the HTTP client code. This will allow us to add a wiremock
test and then verify that we properly encode the URL for userId
lookups. This will then lead to a bug fix where the URL formed
by user names that have spaces in them are not properly URL encoded.