Fix bots not loading by ensuring we reset the game before doing network communications
A call to 'setGameStatus' goes over network and can fail with exception.
If that happens then we will not invoke 'onGameInterrupt' which is
needed for a bot server to be able to start new games or load save games.
Specifically 'onGameInterrupt' will null out a game data which allows
for a game save to be changed and it will re-spawn the thread that
waits for a game to fill up to then start a game. If either of these
do not happen then maps won't change and a bot won't start games.

By re-odering these two calls, we ensure that we'll reset the bot before
any network communication happens.