Lobby: Turn off rate-limiting for bots
Set the endpoints used by bots to 'report-only' for rate limiting.
This will prevent any 429 (rate limit) errors from being thrown
and instead the server will just log when the rate-limit has
been breached but will still service the request.

The bots are not currently set up to handle arbitrary communication
failures and we need to audit them to ensure that they handle
such failures well and do not result in a poison-pill scenario.

This update will let us know where and what breaches rate limits
so we can adjust limits and the code accordingly and gives us
time to harden the bot communication with the lobby.