Use unmodifiable collection in UnitSeperator in Swing Event Thread Task
Fixes #7066 UnitSeparator#categorize:103 - ConcurrentModificationException

This fix avoids concurrent modification to a list by creating a copy
of the list before we operate on the list in the swing even thread.
We go on to fix any other data issues that could change between when
we issue the task to the swing event queue by computing all data
the task needs before creating the task.