Make warnings non-reportable (#7073)
Add a check for logging level in our global logging handler. If the log
level is warning then disable the 'upload button'. Same, if the level
is a warning also add a generic statement giving the user a link to
report the problem.

This update allows us to notify users of general errors that have happened
that are outside of TripleA's control and are something that we can
tell the user how to fix in the warning message. This way we can
use a 'warning' logging to report this to the user without
constructing a full swing dialog with the same text.

Events logged at a 'severe' level will continue to have an 'upload' button
enabled and there is no change to the 'severe' logging behavior.