Add log aggregation to servers
Sets up log aggregation from the application servers to a support server.
An hours worth of logs are dumped from each application using journalctl
and then SCP'd to the support server.

Summary of log aggregation flow:
  - cronjob executes on app server each hour, for each application
    - this creates a log file using journalctl containing the last hours worth of logs
    - the app server uses "logs user" to SCP the log file to aggregration server
    - app server cleans up the log file
  - the log server each hour will zip up log files older than 2 hours
  - the log server each hour will delete log files older than 3 months
  - the log server each hour will delete empty log folders

Of some note, in this update we add 'admin' user to the 'adm' group on
application servers to allow 'admin' user sudo-less journalctl. Without
that you need to run 'sudo' to be able to view logs for arbitrary systemctl
services.

As part of this we update inventory and add a 'support01' server that has
the aggregration server role

In sum, two new roles are added:
- New role: log_aggregation_server
  - sets up a server to receive log files
  - adds a 'logs' user with a public key that other hosts will use when performing
    SCP to copy log files to the host.
  - log files will land in '/home/admin/logs/2020/...'
- New role: log_aggregation_sender:
  - adds the logs user private ssh key to admin user to enable SCP and SSH to the
    log aggregation server
  - adds an hourly cronjob to send last hours logs to the aggregation server
  - adds a 'send_logs' scripts that is invoked by cron that actually sends the log files

The log files that are sent are controlled by a 'group_var' variable: 'journalctl_applications'. Each
entry in the application list will be one item where we will query journalctl (ie: journalctl -u 'app01'),
thereby grabbing the last hours worth of logs and dumping that to a file.