Better fix for leaked DB password, use loop-control
Loop_control can be used to control how the loop iteration label is printed.
When ansible runs a loop, it prints the value of the loop variable.
This is bad for us if the loop variable is a map and contains a password.
This fix removes the skip-tags command that skipped over DB tasks to prevent
password leakage and simply makes explicit which parts of the loop
variable is to be printed (omitting passwords).