Automatically add map-name to error upload reports
Use UIContext.resourceLoader to get map name for error reports.
The upload value for map-name might be incorrect as the game may
not be started, but we'll be guaranteed to have the map name in
error reports and should be able to figure out from a stack trace
if a problem happened in a staging screen or if it was in-game.