Remove 'ignore missing territories block'
This block does not do anything.. We later check if all map
territories are in the keyset that comes from a save game. If
we have extra territories in the save game, removing them is
not going to have an effect when we check that all territories
in the map are in the save game.

IE:

The following contains check will be no different if we decrease the size
of the 'keys' set:
```
      if (!keys.contains(terr.getName())) {
```

It appears an error used to be logged here instead of the 'skip'. It seems
that instead of skipping the iter.remove and surrounding loop could have
been removed in 0eaf35b