Simplify last map update check parsing and set threshold for last check to 7 days
1. Simplify: Rather than storing a formatted string for the last time we checked for
map updates, just store the equivalent epoch milli timestamp.

2. Adjust threshold for checking for maps to be every 7 days rather than once
a month. The once a month made more sense when the map list was hosted on source
forge and the download time was very slow, with a fast download of the map
list we can check more often and thereby get better responsiveness to map
version updates.

Fixes last map-date-check parsing error reported in:
https://github.com/triplea-game/triplea/issues/7014