Ignore failure when retreiving download lengths
We have a fallback for download progress bars to count up the number
of downloaded bytes if we can't print a percentage. If we can't
get the download length, then we can ignore the problem since the progress
bar still works and very notably the download is still in-progress.

This update simply suppresses any errors when requesting the download
length (of map files when downloading them).

Resolves error-report: https://github.com/triplea-game/triplea/issues/7024