Fix: catch errors when launching a server game and abort cleanly
If there are any errors during game load then players will not
be ejected from their seats. In this case when the game stops,
for bots, all seats will still be occupied, the game will
not actually be running, and users joining will get a "game is
launching, try joining later" message. This is a dead-state
for bots as the seats are full, cannot be emptied, and a game
is not actually running. This update resolves that situation
by resetting games if there any errors on load.

Of note, it's critical the game would reach the 'launch' functionality
to clear the 'isLaunching' flag. If we terminate earlier then the
game would be stuck in a 'isLaunching' state.