Simplify PR template, remove 'functional changes' checkbox list
This update removes the checkbox list & simplifies some of
the leading content to be less verbose.

We often skip the checkbox list, hopefully the type of functional
change being submitted is obvious by the commit commentary. If
that is not obvious, then it will not be obvious to maintainers
viewing the code commit history.