Fix setting of database password, use postgres password (#7317)
We need to use the postgres user password when setting
passwords and logging in as the postgres user.

Using 'item.password' would be the database (application)
user and would be incorrect. The task appears to fail
silently when the password is incorrect, the task
was previously not updating passwords as expected.