Increase code climate method argument threshold to 6 arguments or more [skip ci]
Generally, 4 or 5 arguments are too many and we should strongly
consider using a paramter object or other techniques to reduce arg count.
5 arguments seems to be a threshold between when it's bad idea compared
to really beyond the pale. 6 arguments or more probably should always
be behind a parameter object.

This update sets code climate to not generate a warning until
method arguments hit 6 or more, a point where generally we will
want something to be done (reduces false positives)