Update: Change the data structure of map-listing JSON response to account for map skins
This update mostly makes map skins be children of the map element in the map listing JSON
response. Currently we need map downloads (skins included) to be a flat list, so we then
need some client logic to combine map downloads and their map skins into a flat list.

As part of this, we can remove the 'download type' attribute from the JSON response. The
download type is evident from the response structure. One implication is that in the
future we'll remove the 'tool' category and fold that into the 'map' category or
make the 'tool' download availabe via some other means.