Use DropwizardServerExtension to clean up data after test execution
This update adds functionality to the extensions 'afterAll' block
to execute the SQL statements in a test/resources/db-cleanup.sql
file if it exists.

This update allows for most 'cleanBefore' tags to be removed and
removes non-determinism of DbRider tests where certain test
execution orderings can result in a FK constraint violation error.

Motivation & Benefit: DBRider tests will clear tables before inserting
data. One problem is that the table clearing logic is only aware of
the tables in that test. If other tables have a FK to one of these
cleared tables, there will be a constraint violation error. Worse yet,
this situation depends on the ordering of when tests are executed.

By deleting data using SQL from a file that is aware of all tables
and the correct order to delete data, we can execute this to ensure
that when any new tests run that there won't be any foreign key
reference violations when setting up new data.