Update DropWizardExtension: clean after each test & init JDBI once & fail if server already started
1. Re-creating DB connection takes some time, if we already have a DB
connection (ie: jdbi != null), then do not re-created it. This is
a matter of adding an if null check around the JDBI instantiation.

2. Clean DB after each test, this helps avoid FK violations between
tests within a test class. The execution of the clean-up is quite
fast, this makes it easier to write valid tests without slowing
things down noticably.

3. To improve error handling and error messaging, configure the drop wizard
extension server startup (starts server when tests execute) to fail if
the server is already running. Given we no longer manually start a server
in CI, this is okay.